Reslove hard coding fp32 in linear and llama attention

+++ python/llm/src/bigdl/llm/transformers/models/llama.py
@@ -215,7 +215,7 @@ def llama_attention_forward_4_31(
 
     # upcast attention to fp32
     attn_weights = nn.functional.softmax(attn_weights, dim=-1,
-                                         dtype=torch.float32).to(query_states.dtype)
+                                         dtype=attn_weights.dtype).to(query_states.dtype)
     attn_output = torch.matmul(attn_weights, value_states)
 
     attn_output_size = (bsz, self.num_heads, q_len, self.head_dim)
